requirements.txt
requests
beautifulsoup4
opencv-python
dlib
face_recognition
rembg
Pillow
torch
torchvision
